Misplaced modifier
"Wrapped in these firm loli boobies,
I'll rub and scrub your penis...❤*"

"Wrapped in these firm loli boobies" appears to modify/describe "I", because it is the closest thing to that phrase. The succubus isn't what's "wrapped in these firm loli boobies", it's the penis. So I recommend the following restructuring of the sentence:

"I'll rub and scrub your penis,
wrapped in my firm loli boobies...❤*"