<TK201> through <TK206> and Ch8.txt proofreading
<TK202> in qa12.txt corresponds with <TK6> of ch8.txt.
<TK203> in qa12.txt corresponds with <TK7> of ch8.txt.

Ch8.txt line 5080 I changed "when" to "went". I considered whether or not it was pronounced this way since this is a loli, but I don't think it fits her.

<TK204> in qa12.txt corresponds with <TK26> of ch8.txt.

<TK205> in qa12.txt corresponds with <AT26> of ch8.txt.

<TK206> in qa12.txt corresponds with <AT0> of ch8.txt.

Ch8.txt line 1520 should mention "again" or "another" as <AT0> is a variation of <AT26>; <AT0> is if the Drain Imp has already drained someone before.