Lines 16134 - 16779 Copy-Pasta and Proofreading
CM63-2 = CM120

Line 9986 Rewrote the sentence to make better sense.

Line 9986 Replaced "to" with "too" because otherwise it sounds like he wanted to complete a sentence. I'm not sure "too" is needed though, given all he said was a drawn out "気持ちいいぃ・・・❤".

CM78 = CM121
CM78-2 = CM122

Almost all of the text in CM121 and CM122 were identical, so I ran a diffcheck on both, and changed CM122 to match CM121 where appropriate (i.e. wherever the Japanese text was verbatim, which was pretty much everything after the first paragraph of CM122. The first two paragraphs of CM122 was the variation in text you get when you decide to fuck loli Mirage after all).