Rest.txt done, other untranslated lines found
In qa6.txt, lines 215, 767, 768 are untranslated and are in the game, found by Zariel.

In qa12.txt, lines 712, 713, 716, 719, 728-731, 734-735, 8354-8358, 17547-17566, 30022-30027, 41959, 41962-41970, 41973, 41974, 41977, 41980, 41981 are untranslated.
Of these, 728-731, 734, 735, are untranslated and in the game, found by Zariel.

In qa5.txt, lines 1421 and 1422 are untranslated and in the game, found by Zariel.

In rest.txt, finished proofreading everything.