QA13 done
qa14 is completely Japanese, looks like it has the generic healing circle lines and something about waves. Survival mode? Added the generic healing lines in case it comes up later.

Line 732-733. Japanese implies that these are soldiers, not just prisoners, so I changed it a bit.

Line 749-750. Same as lines 784 and 785 in Japanese, copied that translation over this one since it sounded better.