QA13 final, QA9 done
QA13 didn't save all changes for some reason. Remainder of it is here.

qa10 is still in Japanese.

For QA9

OP1 name tags are wrong, even in the original Japanese.
Lines 1439, 1444, 1448, and 1451 in the English file.
Lines 1380, 1384, 1387, and 1390 in the Japanese file.

Lines 1618-1619 sound a bit off, MTL of the Japanese line doesn't make the meaning clear to rewrite it.
The Japanese line is 1545.
Please take a look when you can.